refactored base setup for app
