+++ app/lib/Screens/HomeScreens/HomeTabBar.dart
@@ -1,4 +1,4 @@
-import 'package:effektio/Common+Store/Colors.dart';
+import 'package:effektio/common/store/Colors.dart';
 import 'package:effektio/Screens/HomeScreens/News.dart';
 import 'package:effektio/Screens/HomeScreens/Notification.dart';
 import 'package:flutter/material.dart';
