+++ app/lib/Screens/HomeScreens/News.dart
@@ -1,10 +1,10 @@
 // ignore_for_file: prefer_const_constructors
 
-import 'package:effektio/Common+Store/Colors.dart';
+import 'package:effektio/common/store/Colors.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
-import 'package:effektio/Common+Widget/SideMenu.dart';
+import 'package:effektio/common/widget/SideMenu.dart';
 
 class NewsScreen extends StatefulWidget {
   const NewsScreen({Key? key}) : super(key: key);
