+++ app/lib/Screens/SideMenuScreens/Gallery.dart
@@ -1,6 +1,6 @@
-import 'package:effektio/Common+Store/Colors.dart';
-import 'package:effektio/Common+Widget/AppCommon.dart';
-import 'package:effektio/Common+Widget/Feed.dart';
+import 'package:effektio/common/store/Colors.dart';
+import 'package:effektio/common/widget/AppCommon.dart';
+import 'package:effektio/common/widget/Feed.dart';
 import 'package:flutter/material.dart';
 
 class GalleryScreen extends StatefulWidget {
